Arrow task and exception
