+++ src/main/java/com/snowleopard1863/APTurrets/exception/ArrowLaunchException.java
@@ -1,0 +1,7 @@
+package com.snowleopard1863.APTurrets.exception;
+
+public class ArrowLaunchException extends RuntimeException {
+    public ArrowLaunchException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
